+++ apps/server/src/meeting-session/dto/create-meeting-session.dto.ts
@@ -2,20 +2,30 @@ import { IsDateString, IsNotEmpty, IsString } from "class-validator"
 
 export class CreateMeetingSessionDto {
     @IsString()
-    @IsNotEmpty()
+    @IsNotEmpty({
+        message: "Nama Sesi Rapat Tidak Boleh Kosong",
+    })
     name: string
     @IsDateString({
         strict: true,
     })
-    @IsNotEmpty()
+    @IsNotEmpty({
+        message: "Tanggal Sesi Rapat Tidak Boleh Kosong"
+    })
     date: string
     @IsString()
-    @IsNotEmpty()
+    @IsNotEmpty({
+        message: "Tempat Sesi Rapat Tidak Boleh Kosong"
+    })
     location: string
     @IsString()
-    @IsNotEmpty()
+    @IsNotEmpty({
+        message: "Agenda Sesi Rapat Tidak Boleh Kosong"
+    })
     agenda: string
     @IsString()
-    @IsNotEmpty()
+    @IsNotEmpty({
+        message: "Waktu Sesi Rapat Tidak Boleh Kosong"
+    })
     time: string
 }