+++ apps/server/src/public/public.module.ts
@@ -1,0 +1,12 @@
+import { Module } from "@nestjs/common";
+import { PublicController } from "./public.controller";
+import { PublicService } from "./public.service";
+
+@Module({
+    imports: [],
+    providers: [PublicService],
+    controllers: [PublicController]
+})
+export class PublicModule {
+
+}