+++ apps/server/src/sessions/sessions.controller.ts
@@ -8,11 +8,11 @@ import { SessionRoleType } from '@prisma/client';
 
 @Controller('sessions')
 @ApiTags("Gateway")
+@UseGuards(AccessTokenGuard)
 export class SessionsController {
   constructor(private readonly sessionsService: SessionsService) { }
 
   @Post()
-  @UseGuards(AccessTokenGuard)
   create(@Body() createSessionDto: CreateSessionDto) {
     return this.sessionsService.create(createSessionDto);
   }
@@ -28,19 +28,16 @@ export class SessionsController {
   }
 
   @Get(':id')
-  @UseGuards(AccessTokenGuard)
   findOne(@Param('id') id: string) {
     return this.sessionsService.findOne(+id);
   }
 
   @Patch(':id')
-  @UseGuards(AccessTokenGuard)
   update(@Param('id') id: string, @Body() updateSessionDto: UpdateSessionDto) {
     return this.sessionsService.update(+id, updateSessionDto);
   }
 
   @Delete(':id')
-  @UseGuards(AccessTokenGuard)
   remove(@Param('id') id: string) {
     return this.sessionsService.remove(+id);
   }