+++ apps/server/src/sessions/sessions.service.ts
@@ -138,7 +138,6 @@ export class SessionsService {
             location: true,
             role: true,
             status: true,
-            // todo : needed to remove
             token: true
           }
         }